add download file

+++ app/controllers/internal_api/datasets_controller.rb
@@ -2,7 +2,8 @@ class InternalApi::DatasetsController < InternalApi::ApplicationController
   before_action :authenticate_user, except: :index
 
   def index
-    res_body = Starhub.api.get_datasets(params[:search],
+    res_body = Starhub.api.get_datasets(current_user&.name,
+                                        params[:search],
                                         params[:sort],
                                         params[:task_tag],
                                         params[:framework_tag],
@@ -32,6 +33,11 @@ class InternalApi::DatasetsController < InternalApi::ApplicationController
 
     unless @dataset
       return render json: { message: "未找到对应数据集" }, status: 404
+    end
+
+    unless current_user.can_manage?(@dataset)
+      render json: { message: '无权限' }, status: :unauthorized
+      return
     end
 
     if @dataset.destroy
