+++ app/controllers/internal_api/models_controller.rb
@@ -2,7 +2,8 @@ class InternalApi::ModelsController < InternalApi::ApplicationController
   before_action :authenticate_user, except: :index
 
   def index
-    res_body = Starhub.api.get_models(params[:search],
+    res_body = Starhub.api.get_models(current_user&.name,
+                                      params[:search],
                                       params[:sort],
                                       params[:task_tag],
                                       params[:framework_tag],
@@ -34,6 +35,11 @@ class InternalApi::ModelsController < InternalApi::ApplicationController
       return render json: { message: "未找到对应模型" }, status: 404
     end
 
+    unless current_user.can_manage?(@model)
+      render json: { message: '无权限' }, status: :unauthorized
+      return
+    end
+
     if params[:private].to_s == 'true'
       @model.visibility = 'private'
     else
@@ -53,6 +59,11 @@ class InternalApi::ModelsController < InternalApi::ApplicationController
 
     unless @model
       return render json: { message: "未找到对应模型" }, status: 404
+    end
+
+    unless current_user.can_manage?(@model)
+      render json: { message: '无权限' }, status: :unauthorized
+      return
     end
 
     if @model.destroy
